The rest API functions were defined two times in "helper.py" and "__init__,py" as well. That caused the confusion mentioned in issue #294. Now all the helper functions are moved to "helper.py".
